+++ src/Features/Core/Portable/InitializeParameter/AbstractInitializeMemberFromParameterCodeRefactoringProviderMemberCreation.cs
@@ -6,7 +6,6 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -76,7 +75,7 @@ internal abstract partial class AbstractInitializeMemberFromParameterCodeRefacto
 
         // See if we're already assigning this parameter to a field/property in this type. If so, there's nothing
         // more for us to do.
-        var assignmentStatement = TryFindFieldOrPropertyAssignmentStatement(parameter, blockStatement);
+        var assignmentStatement = InitializeParameterHelpersCore.TryFindFieldOrPropertyAssignmentStatement(parameter, blockStatement);
         if (assignmentStatement != null)
             return [];
 
@@ -227,7 +226,7 @@ internal abstract partial class AbstractInitializeMemberFromParameterCodeRefacto
             if (parameterNameParts.BaseName == "")
                 continue;
 
-            var assignmentOp = TryFindFieldOrPropertyAssignmentStatement(parameter, blockStatement);
+            var assignmentOp = InitializeParameterHelpersCore.TryFindFieldOrPropertyAssignmentStatement(parameter, blockStatement);
             if (assignmentOp != null)
                 continue;
 
@@ -478,7 +477,7 @@ internal abstract partial class AbstractInitializeMemberFromParameterCodeRefacto
                 });
         }
 
-        AddAssignment(constructorDeclaration, blockStatement, parameter, fieldOrProperty, editor);
+        AddAssignment<TStatementSyntax>(constructorDeclaration, blockStatement, parameter, fieldOrProperty, editor);
 
         // If the user had a property that has 'throw NotImplementedException' in it, then remove those throws.
         var currentSolution = document.Project.Solution;
@@ -507,36 +506,6 @@ internal abstract partial class AbstractInitializeMemberFromParameterCodeRefacto
         }
 
         return currentSolution.WithDocumentSyntaxRoot(document.Id, editor.GetChangedRoot());
-    }
-
-    private void AddAssignment(
-        SyntaxNode constructorDeclaration,
-        IBlockOperation? blockStatement,
-        IParameterSymbol parameter,
-        ISymbol fieldOrProperty,
-        SyntaxEditor editor)
-    {
-        // First see if the user has `(_x, y) = (x, y);` and attempt to update that. 
-        if (TryUpdateTupleAssignment(blockStatement, parameter, fieldOrProperty, editor))
-            return;
-
-        var generator = editor.Generator;
-
-        // Now that we've added any potential members, create an assignment between it
-        // and the parameter.
-        var initializationStatement = (TStatementSyntax)generator.ExpressionStatement(
-            generator.AssignmentStatement(
-                generator.MemberAccessExpression(
-                    generator.ThisExpression(),
-                    generator.IdentifierName(fieldOrProperty.Name)),
-                generator.IdentifierName(parameter.Name)));
-
-        // Attempt to place the initialization in a good location in the constructor
-        // We'll want to keep initialization statements in the same order as we see
-        // parameters for the constructor.
-        var statementToAddAfter = TryGetStatementToAddInitializationAfter(parameter, blockStatement);
-
-        InsertStatement(editor, constructorDeclaration, returnsVoid: true, statementToAddAfter, initializationStatement);
     }
 
     private static CodeGenerationContext GetAddContext<TSymbol>(
@@ -572,63 +541,6 @@ internal abstract partial class AbstractInitializeMemberFromParameterCodeRefacto
         }
 
         return CodeGenerationContext.Default;
-    }
-
-    private SyntaxNode? TryGetStatementToAddInitializationAfter(
-        IParameterSymbol parameter, IBlockOperation? blockStatement)
-    {
-        // look for an existing assignment for a parameter that comes before/after us.
-        // If we find one, we'll add ourselves before/after that parameter check.
-        foreach (var (sibling, before) in GetSiblingParameters(parameter))
-        {
-            var statement = TryFindFieldOrPropertyAssignmentStatement(sibling, blockStatement);
-            if (statement != null)
-            {
-                if (before)
-                {
-                    return statement.Syntax;
-                }
-                else
-                {
-                    var statementIndex = blockStatement!.Operations.IndexOf(statement);
-                    return statementIndex > 0 && blockStatement.Operations[statementIndex - 1] is { IsImplicit: false, Syntax: var priorSyntax }
-                        ? priorSyntax
-                        : null;
-                }
-            }
-        }
-
-        // We couldn't find a reasonable location for the new initialization statement.
-        // Just place ourselves after the last statement in the constructor.
-        return TryGetLastStatement(blockStatement);
-    }
-
-    private static IOperation? TryFindFieldOrPropertyAssignmentStatement(IParameterSymbol parameter, IBlockOperation? blockStatement)
-        => TryFindFieldOrPropertyAssignmentStatement(parameter, blockStatement, out _);
-
-    protected static bool TryGetPartsOfTupleAssignmentOperation(
-        IOperation operation,
-        [NotNullWhen(true)] out ITupleOperation? targetTuple,
-        [NotNullWhen(true)] out ITupleOperation? valueTuple)
-    {
-        if (operation is IExpressionStatementOperation
-            {
-                Operation: IDeconstructionAssignmentOperation
-                {
-                    Target: ITupleOperation targetTupleTemp,
-                    Value: IConversionOperation { Operand: ITupleOperation valueTupleTemp },
-                }
-            } &&
-            targetTupleTemp.Elements.Length == valueTupleTemp.Elements.Length)
-        {
-            targetTuple = targetTupleTemp;
-            valueTuple = valueTupleTemp;
-            return true;
-        }
-
-        targetTuple = null;
-        valueTuple = null;
-        return false;
     }
 
     private static IOperation? TryFindFieldOrPropertyAssignmentStatement(
