+++ src/Features/Core/Portable/InitializeParameter/AbstractInitializeParameterCodeRefactoringProvider.cs
@@ -20,6 +20,8 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.InitializeParameter;
 
+using static InitializeParameterHelpersCore;
+
 internal abstract partial class AbstractInitializeParameterCodeRefactoringProvider<
     TTypeDeclarationSyntax,
     TParameterSyntax,
@@ -194,47 +196,6 @@ internal abstract partial class AbstractInitializeParameterCodeRefactoringProvid
                 return true;
         }
 
-        return false;
-    }
-
-    protected static bool IsFieldOrPropertyAssignment(IOperation statement, INamedTypeSymbol containingType, [NotNullWhen(true)] out IAssignmentOperation? assignmentExpression)
-        => IsFieldOrPropertyAssignment(statement, containingType, out assignmentExpression, out _);
-
-    protected static bool IsFieldOrPropertyAssignment(
-        IOperation statement, INamedTypeSymbol containingType,
-        [NotNullWhen(true)] out IAssignmentOperation? assignmentExpression,
-        [NotNullWhen(true)] out ISymbol? fieldOrProperty)
-    {
-        if (statement is IExpressionStatementOperation expressionStatement &&
-            expressionStatement.Operation is IAssignmentOperation assignment)
-        {
-            assignmentExpression = assignment;
-            return IsFieldOrPropertyReference(assignmentExpression.Target, containingType, out fieldOrProperty);
-        }
-
-        fieldOrProperty = null;
-        assignmentExpression = null;
-        return false;
-    }
-
-    protected static bool IsFieldOrPropertyReference(IOperation operation, INamedTypeSymbol containingType)
-        => IsFieldOrPropertyAssignment(operation, containingType, out _);
-
-    protected static bool IsFieldOrPropertyReference(
-        IOperation? operation, INamedTypeSymbol containingType,
-        [NotNullWhen(true)] out ISymbol? fieldOrProperty)
-    {
-        if (operation is IMemberReferenceOperation memberReference &&
-            memberReference.Member.ContainingType.Equals(containingType))
-        {
-            if (memberReference.Member is IFieldSymbol or IPropertySymbol)
-            {
-                fieldOrProperty = memberReference.Member;
-                return true;
-            }
-        }
-
-        fieldOrProperty = null;
         return false;
     }
 }
